+++ src/App.tsx
@@ -11,6 +11,7 @@ import {
   PatientAllergies,
   PatientConditions,
   PatientConditionsOutside,
+  PatientDiagnosticReports,
   PatientDocuments,
   PatientImmunizations,
   PatientMedications,
@@ -78,6 +79,11 @@ const components: DemoComponent[] = [
     title: "Patient Conditions Outside",
   },
   {
+    name: "diagnostic-reports",
+    render: () => <PatientDiagnosticReports />,
+    title: "Patient Diagnostic Reports",
+  },
+  {
     name: "documents",
     render: () => <PatientDocuments />,
     title: "Patient Documents",
@@ -111,16 +117,16 @@ const components: DemoComponent[] = [
         conditionsOutsideProps={{
           hideRequestRecords: true,
         }}
+        conditionsAllProps={{ onlyAllowAddOutsideConditions: true }}
         includePatientDemographicsForm={false}
         resources={[
           "allergies",
-          "conditions",
-          "conditions-outside",
+          "conditions-all",
           "documents",
           "immunizations",
           "care-team",
-          "medications",
-          "medications-outside",
+          "medications-all",
+          "diagnostic-reports",
         ]}
         title="ZAP"
       />