+++ src/components/content/patient-history/patient-history-table.stories.tsx
@@ -2,7 +2,6 @@ import { Meta, StoryObj } from "@storybook/react";
 import { PatientHistoryTable } from "./patient-history-table";
 import { setupPatientHistoryMocks } from "./story-helpers/mocks/requests";
 import { FAKE_AUTH, FAKE_BUILDER_ID, FAKE_PATIENT_UPID } from "../story-helpers/ids";
-import { PatientObservationsProps as Props } from "@/components/content/observations/patient-observations";
 import { CTWProvider } from "@/components/core/providers/ctw-provider";
 import { PatientProvider } from "@/components/core/providers/patient-provider";
 import { SYSTEM_ZUS_UNIVERSAL_ID } from "@/fhir/system-urls";
@@ -20,8 +19,8 @@ export default {
       </CTWProvider>
     ),
   ],
-} as Meta<Props>;
+} as Meta;
 
-export const Basic: StoryObj<Props> = {
+export const Basic: StoryObj = {
   ...setupPatientHistoryMocks(),
 };